requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
pydantic==2.11.7
langchain>=0.3.0,<0.4.0
langchain-ollama>=0.3.0,<0.4.0
langchain-chroma>=0.2.0,<0.3.0
langchain-text-splitters>=0.3.0,<0.4.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
